requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
google-generativeai
